backend/requirements.txt
asgiref==3.11.1
Django==4.2.30
django-cors-headers==4.9.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.8
PyJWT==2.12.1
python-decouple==3.8
pytz==2026.1.post1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.2
